src/main/kotlin/no/nav/personopplysninger/personalia/transformer/GeografiskTilknytningTransformer.kt
package no.nav.personopplysninger.personalia.transformer

import no.nav.personopplysninger.personalia.dto.PersonaliaKodeverk
import no.nav.personopplysninger.personalia.dto.outbound.GeografiskTilknytning


object GeografiskTilknytningTransformer {

    fun toOutbound(inbound: no.nav.pdl.generated.dto.hentpersonquery.GeografiskTilknytning, kodeverk: PersonaliaKodeverk): GeografiskTilknytning = GeografiskTilknytning(
            bydel = inbound.gtBydel,
            kommune = inbound.gtKommune,
            land = kodeverk.gtLandterm
    )

}